changes 1. Seperated the user, user_reques and user_response models
